change: Submission Entity では Status を判定しない * 代わりに Entity に `reviewed` と `finished` を追加 * 完全な `Status` を判定するには、Judge からも情報を取得する必要があり、負荷がかかる * UseCase に `Submission` と `TestResult` (Judge) から `Status` を取得するコードを追加
