+++ src/domain/entities/submission.ts
@@ -1,17 +1,4 @@
-import { z } from 'zod';
 import { SupportedLanguage } from './supported-language';
-
-export const statusEnum = [
-  'PENDING',
-  'FAILED',
-  'IN_REVIEW',
-  'REVIEWED',
-  'FINISHED',
-] as const satisfies string[];
-
-/** Zod */
-export const status = z.enum(statusEnum);
-export type Status = z.infer<typeof status>;
 
 export class Submission {
   constructor(
@@ -20,7 +7,8 @@ export class Submission {
     public readonly problemId: string,
     public readonly language: SupportedLanguage,
     public readonly codeText: string,
-    public readonly status: Status,
+    public readonly reviewed: boolean,
+    public readonly finished: boolean,
     public readonly testResultId: string,
     public readonly createdAt: Date,
     public readonly updatedAt: Date
